+++ Source/Scene/GoogleEarthEnterpriseImageryProvider.js
@@ -228,7 +228,7 @@ function GoogleEarthEnterpriseImageryProvider(options) {
     .then(function (result) {
       if (!metadata.imageryPresent) {
         const e = new RuntimeError(
-          "The server " + metadata.url + " doesn't have imagery"
+          `The server ${metadata.url} doesn't have imagery`
         );
         metadataError = TileProviderError.handleError(
           metadataError,
@@ -663,7 +663,7 @@ function buildImageResource(imageryProvider, info, x, y, level, request) {
   version = defined(version) && version > 0 ? version : 1;
 
   return imageryProvider._metadata.resource.getDerivedResource({
-    url: "flatfile?f1-0" + quadKey + "-i." + version.toString(),
+    url: `flatfile?f1-0${quadKey}-i.${version.toString()}`,
     request: request,
   });
 }