+++ Source/Scene/ShadowMapShader.js
@@ -12,16 +12,7 @@ ShadowMapShader.getShadowCastShaderKeyword = function (
   usesDepthTexture,
   isOpaque
 ) {
-  return (
-    "castShadow " +
-    isPointLight +
-    " " +
-    isTerrain +
-    " " +
-    usesDepthTexture +
-    " " +
-    isOpaque
-  );
+  return `castShadow ${isPointLight} ${isTerrain} ${usesDepthTexture} ${isOpaque}`;
 };
 
 ShadowMapShader.createShadowCastVertexShader = function (
@@ -107,15 +98,13 @@ ShadowMapShader.createShadowCastFragmentShader = function (
 
   if (isPointLight) {
     fsSource +=
-      "    float distance = length(" +
-      positionVaryingName +
-      "); \n" +
-      "    if (distance >= shadowMap_lightPositionEC.w) \n" +
-      "    { \n" +
-      "        discard; \n" +
-      "    } \n" +
-      "    distance /= shadowMap_lightPositionEC.w; // radius \n" +
-      "    gl_FragColor = czm_packDepth(distance); \n";
+      `    float distance = length(${positionVaryingName}); \n` +
+      `    if (distance >= shadowMap_lightPositionEC.w) \n` +
+      `    { \n` +
+      `        discard; \n` +
+      `    } \n` +
+      `    distance /= shadowMap_lightPositionEC.w; // radius \n` +
+      `    gl_FragColor = czm_packDepth(distance); \n`;
   } else if (usesDepthTexture) {
     fsSource += "    gl_FragColor = vec4(1.0); \n";
   } else {
@@ -146,19 +135,7 @@ ShadowMapShader.getShadowReceiveShaderKeyword = function (
   const debugCascadeColors = shadowMap.debugCascadeColors;
   const softShadows = shadowMap.softShadows;
 
-  return (
-    "receiveShadow " +
-    usesDepthTexture +
-    polygonOffsetSupported +
-    isPointLight +
-    isSpotLight +
-    hasCascades +
-    debugCascadeColors +
-    softShadows +
-    castShadows +
-    isTerrain +
-    hasTerrainNormal
-  );
+  return `receiveShadow ${usesDepthTexture}${polygonOffsetSupported}${isPointLight}${isSpotLight}${hasCascades}${debugCascadeColors}${softShadows}${castShadows}${isTerrain}${hasTerrainNormal}`;
 };
 
 ShadowMapShader.createShadowReceiveVertexShader = function (
@@ -260,7 +237,7 @@ ShadowMapShader.createShadowReceiveFragmentShader = function (
 
   let returnPositionEC;
   if (hasPositionVarying) {
-    returnPositionEC = "    return vec4(" + positionVaryingName + ", 1.0); \n";
+    returnPositionEC = `    return vec4(${positionVaryingName}, 1.0); \n`;
   } else {
     returnPositionEC =
       "#ifndef LOG_DEPTH \n" +
@@ -271,34 +248,34 @@ ShadowMapShader.createShadowReceiveFragmentShader = function (
   }
 
   fsSource +=
-    "uniform mat4 shadowMap_matrix; \n" +
-    "uniform vec3 shadowMap_lightDirectionEC; \n" +
-    "uniform vec4 shadowMap_lightPositionEC; \n" +
-    "uniform vec4 shadowMap_normalOffsetScaleDistanceMaxDistanceAndDarkness; \n" +
-    "uniform vec4 shadowMap_texelSizeDepthBiasAndNormalShadingSmooth; \n" +
-    "#ifdef LOG_DEPTH \n" +
-    "varying vec3 v_logPositionEC; \n" +
-    "#endif \n" +
-    "vec4 getPositionEC() \n" +
-    "{ \n" +
-    returnPositionEC +
-    "} \n" +
-    "vec3 getNormalEC() \n" +
-    "{ \n" +
-    (hasNormalVarying
-      ? "    return normalize(" + normalVaryingName + "); \n"
-      : "    return vec3(1.0); \n") +
-    "} \n" +
+    `${
+      "uniform mat4 shadowMap_matrix; \n" +
+      "uniform vec3 shadowMap_lightDirectionEC; \n" +
+      "uniform vec4 shadowMap_lightPositionEC; \n" +
+      "uniform vec4 shadowMap_normalOffsetScaleDistanceMaxDistanceAndDarkness; \n" +
+      "uniform vec4 shadowMap_texelSizeDepthBiasAndNormalShadingSmooth; \n" +
+      "#ifdef LOG_DEPTH \n" +
+      "varying vec3 v_logPositionEC; \n" +
+      "#endif \n" +
+      "vec4 getPositionEC() \n" +
+      "{ \n"
+    }${returnPositionEC}} \n` +
+    `vec3 getNormalEC() \n` +
+    `{ \n${
+      hasNormalVarying
+        ? `    return normalize(${normalVaryingName}); \n`
+        : "    return vec3(1.0); \n"
+    }} \n` +
     // Offset the shadow position in the direction of the normal for perpendicular and back faces
-    "void applyNormalOffset(inout vec4 positionEC, vec3 normalEC, float nDotL) \n" +
-    "{ \n" +
-    (bias.normalOffset && hasNormalVarying
-      ? "    float normalOffset = shadowMap_normalOffsetScaleDistanceMaxDistanceAndDarkness.x; \n" +
-        "    float normalOffsetScale = 1.0 - nDotL; \n" +
-        "    vec3 offset = normalOffset * normalOffsetScale * normalEC; \n" +
-        "    positionEC.xyz += offset; \n"
-      : "") +
-    "} \n";
+    `void applyNormalOffset(inout vec4 positionEC, vec3 normalEC, float nDotL) \n` +
+    `{ \n${
+      bias.normalOffset && hasNormalVarying
+        ? "    float normalOffset = shadowMap_normalOffsetScaleDistanceMaxDistanceAndDarkness.x; \n" +
+          "    float normalOffsetScale = 1.0 - nDotL; \n" +
+          "    vec3 offset = normalOffset * normalOffsetScale * normalEC; \n" +
+          "    positionEC.xyz += offset; \n"
+        : ""
+    }} \n`;
 
   fsSource +=
     "void main() \n" +
@@ -359,7 +336,7 @@ ShadowMapShader.createShadowReceiveFragmentShader = function (
       "    shadowParameters.nDotL = nDotL; \n" +
       "    float visibility = czm_shadowVisibility(shadowMap_texture, shadowParameters); \n";
   } else if (hasCascades) {
-    fsSource +=
+    fsSource += `${
       "    float maxDepth = shadowMap_cascadeSplits[1].w; \n" +
       "    // Stop early if the eye depth exceeds the last cascade \n" +
       "    if (depth > maxDepth) \n" +
@@ -381,11 +358,13 @@ ShadowMapShader.createShadowReceiveFragmentShader = function (
       "    // Fade out shadows that are far away \n" +
       "    float shadowMapMaximumDistance = shadowMap_normalOffsetScaleDistanceMaxDistanceAndDarkness.z; \n" +
       "    float fade = max((depth - shadowMapMaximumDistance * 0.8) / (shadowMapMaximumDistance * 0.2), 0.0); \n" +
-      "    visibility = mix(visibility, 1.0, fade); \n" +
-      (debugCascadeColors
+      "    visibility = mix(visibility, 1.0, fade); \n"
+    }${
+      debugCascadeColors
         ? "    // Draw cascade colors for debugging \n" +
           "    gl_FragColor *= czm_cascadeColor(weights); \n"
-        : "");
+        : ""
+    }`;
   } else {
     fsSource +=
       "    float nDotL = clamp(dot(normalEC, shadowMap_lightDirectionEC), 0.0, 1.0); \n" +