+++ Specs/ThirdParty/whenSpec.js
@@ -113,7 +113,7 @@ describe("ThirdParty/when", function () {
   it("chains failure", function () {
     promise
       .then(undefined, function (e) {
-        return when.reject("important" + e);
+        return when.reject(`important${e}`);
       })
       .then(captureValue, captureError);